Code cleanup, refactoring and documentation.
