+++ app/src/main/java/com/physical_web/cms/physicalwebcms/MainActivity.java
@@ -9,17 +9,11 @@ import android.widget.TextView;
 
 public class MainActivity extends AppCompatActivity {
     public static final String TAG = "Physical Web CMS";
-    
-    BeaconDBManager beaconDBManager;
-    BluetoothManager bluetoothManager;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        // open database
-        //beaconDBManager = new BeaconDBManager(this);
 
         Intent entrollmentIntent = new Intent(this, EnrollmentActivity.class);
         startActivity(entrollmentIntent);
@@ -40,7 +34,8 @@ public class MainActivity extends AppCompatActivity {
         SetupManager setupManager = new SetupManager(this);
         setupManager.checkRequirements();
 
-        beaconDBManager = new BeaconDBManager(this);
+        // TODO temp showcase
+        BeaconDBManager beaconDBManager = new BeaconDBManager(this);
         ((TextView)findViewById(R.id.helloWorld)).setText("Enrolled beacons:\n " + beaconDBManager.getAllBeacons().toString());
         beaconDBManager.close();
     }
