+++ app/src/main/java/com/physical_web/cms/physicalwebcms/WelcomeActivity.java
@@ -62,3 +62,38 @@ public class WelcomeActivity extends AppCompatActivity {
         }
     }
 }
+
+class SetupManager {
+    private SharedPreferences sharedPreferences;
+    private Context context;
+
+    public SetupManager(Context activityContext) {
+        this.context = activityContext;
+        sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context);
+    }
+
+    public void checkRequirements() {
+        if(isFirstRun()) {
+            Intent welcomeIntent = new Intent(this.context, WelcomeActivity.class);
+            context.startActivity(welcomeIntent);
+        } else {
+            // TODO: check if connection lost while user is in app, warn them
+        }
+    }
+
+    // returns whether the app has been run before. will return false if the WelcomeActivity
+    // runs once successfully
+    private Boolean isFirstRun() {
+        // if variable doesn't exist in sharedPreferences, returns true (vacuous truth)
+        return sharedPreferences.getBoolean("drive-setup-completed", true);
+    }
+
+    // check if we are connected to a network with internet access.
+    public static Boolean networkIsConnected(Context context) {
+        ConnectivityManager cm = (ConnectivityManager)context
+                .getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo activeNetwork = cm.getActiveNetworkInfo();
+
+        return activeNetwork != null && activeNetwork.isConnectedOrConnecting();
+    }
+}
